1.3.1 start; fixed bug where lessons where deleted even though user clicked undo

+++ app/src/main/java/de/oerntec/votenote/SubjectManagerStuff/SubjectManagementActivity.java
@@ -69,6 +69,7 @@ public class SubjectManagementActivity extends AppCompatActivity implements Swip
      * the subject adapter used to fill the list
      */
     private static SubjectAdapter mSubjectAdapter;
+    public boolean abortLessonDelete = false;
     /**
      * whenever a dialog is opened, we save its position in the recyclerview in here. while it is
      * quite ugly, it should work without problems since the only use is from the delete button in
@@ -175,6 +176,8 @@ public class SubjectManagementActivity extends AppCompatActivity implements Swip
         //save the old position
         positionOfSubjectToBeDeleted = undoBarRecyclerViewPosition;
         //TODO: less fucked up way to save the positions in the recyclerview; currently done via 2 class variables :/
+        //reset deletion abort bool so its deleted if no action is taken
+        abortLessonDelete = false;
         //make snackbar
         Snackbar
                 .make(findViewById(R.id.subject_manager_coordinator_layout), getString(R.string.undobar_deleted), Snackbar.LENGTH_LONG)
@@ -189,7 +192,9 @@ public class SubjectManagementActivity extends AppCompatActivity implements Swip
         new Handler().postDelayed(new Runnable() {
             @Override
             public void run() {
-                DBLessons.getInstance().deleteAllEntriesForGroup(subjectId);
+                //check whether "undo" was tapped
+                if (!abortLessonDelete)
+                    DBLessons.getInstance().deleteAllEntriesForGroup(subjectId);
             }
         }, Snackbar.LENGTH_LONG + 50);
     }
@@ -198,6 +203,7 @@ public class SubjectManagementActivity extends AppCompatActivity implements Swip
     private void onUndo() {
         if (subjectToBeDeleted != null)
             mSubjectAdapter.addSubject(subjectToBeDeleted, positionOfSubjectToBeDeleted);
+        abortLessonDelete = true;
         subjectToBeDeleted = null;
         positionOfSubjectToBeDeleted = -1;
     }
